modulo-business-analytics-and-machine-learning/34-Cap17/pyproject.toml
[project]
name = "34-cap17"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.16",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "banks==2.1.1",
    "beautifulsoup4==4.13.3",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "dirtyjson==1.0.8",
    "docx2txt==0.9",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "frozenlist==1.5.0",
    "fsspec==2025.3.2",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "greenlet==3.1.1",
    "griffe==1.7.2",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "huggingface-hub==0.30.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "langchain-core==0.3.50",
    "langchain-huggingface==0.1.2",
    "langsmith==0.3.23",
    "llama-index-core==0.12.28",
    "llama-index-embeddings-huggingface==0.5.2",
    "llama-index-embeddings-langchain==0.3.0",
    "llama-index-llms-ollama==0.5.4",
    "llama-index-readers-file==0.4.7",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mpmath==1.3.0",
    "multidict==6.2.0",
    "mypy-extensions==1.0.0",
    "narwhals==1.33.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.1",
    "numpy==2.2.4",
    "ollama==0.4.7",
    "orjson==3.10.16",
    "packaging==24.2",
    "pandas==2.2.3",
    "pillow==11.1.0",
    "platformdirs==4.3.7",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "pyarrow==19.0.1",
    "pydantic==2.11.1",
    "pydantic-core==2.33.0",
    "pydeck==0.9.1",
    "pypdf==5.4.0",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.24.0",
    "safetensors==0.5.3",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentence-transformers==4.0.1",
    "setuptools==75.8.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.40",
    "streamlit==1.44.1",
    "striprtf==0.0.26",
    "sympy==1.13.1",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "toml==0.10.2",
    "torch==2.6.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "transformers==4.50.3",
    "typing-extensions==4.13.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "watchdog==6.0.0",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zstandard==0.23.0",
]
